+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
 
 ## 1.0.1 [UNRELEASED]
+### Added
+- `bridge.client.openExternalUrl()` for opening URLs with the default browser
+- A link to the OSC reference within settings
+- A new play mode for groups, Random, which will trigger a random child on play
+### Changed
+- Added 'Stop' as an option when setting default on end actions in settings
 ### Fixed
 - Added a check for undefined items in item.change in the type plugin
 - Regular Electron accelerators now work even though a widget isn't focused (CMD+Q e.t.c.)
+- Icons for rundown items became misaligned when the items grew in height
 
 ## 1.0.0
 ### Added